# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/banner": {
      "post": {
        "operationId": "BannerController_createBanner",
        "summary": "Create new banner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        },
        "tags": [
          "banners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BannerController_getAllBanner",
        "summary": "Get all banner",
        "parameters": [
          {
            "name": "statusId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DESC",
                "ASC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        },
        "tags": [
          "banners"
        ]
      }
    },
    "/banner/{id}": {
      "put": {
        "operationId": "BannerController_updateBanner",
        "summary": "Update banner by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        },
        "tags": [
          "banners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "BannerController_deleteBanner",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "banners"
        ]
      }
    },
...
    "/blog": {
      "post": {
        "operationId": "BlogController_createBlog",
        "summary": "Create new Blog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "blog information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCommonDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized!"
          },
          "403": {
            "description": "Somethings went wrong"
          }
        },
        "tags": [
          "blogs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BlogController_searchBlogs",
        "summary": "Search blog by query",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "size",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "statusId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "subjectId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "notDel",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        },
        "tags": [
          "blogs"
        ]
      }
    },
    "/blog/{id}": {
      "put": {
        "operationId": "BlogController_updateBlog",
        "summary": "Updateblog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "blogs"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BlogController_getBlogDetail",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "blogs"
        ]
      },
      "delete": {
        "operationId": "BlogController_deleteBlog",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "blogs"
        ]
      }
    },
...
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AllcodeDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "parentCode": {
            "type": "string"
          },
          "hexCode": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "value",
          "code",
          "parentCode",
          "hexCode"
        ]
      },
...
      "BannerDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "name",
          "statusId",
          "image"
        ]
      },
      "ResponseCommonDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Banner": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AllcodeDto"
          },
          "image": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "description",
          "name",
          "statusId",
          "status",
          "image",
          "createdAt",
          "updatedAt"
        ]
      },
      "BannerApiResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Banner"
            }
          },
          "meta": {
            "type": "object",
            "example": {
              "current": 1,
              "size": 10,
              "totalItems": 100
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "BlogDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "subjectId": {
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contentMarkdown": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "title"
        ]
      },
      "Blog": {
        "type": "object",
        "properties": {}
      },
      "BlogApiResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Blog"
            }
          },
          "meta": {
            "type": "object",
            "example": {
              "current": 1,
              "size": 10,
              "totalItems": {
                "items": 4
              }
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}